requirements.txt
flask>=2.0.3
flask-compress>=1.10.1
flask_login>=0.6.0
gunicorn>=20.1.0
google-cloud-storage>=2.1.0
google-api-python-client>=2.86.0
# these stubs slow mypy down a lot
# google-api-python-client-stubs>=1.17.0
numpy>=1.24.1
pysam>=0.20.0
rauth>=0.7.3
scipy>=1.10.1
types-requests>=2.28
types-setuptools>=67.8.0
types-Flask>=1.1.6
typing_extensions>=4.5.0
